[main] Move database connect after argparse
Accessing the database requires switching to the GitLab postgres user,
which requires root. We should connect to the database as late as
possible.

We had been resolving projects directly in argparse, which was very
clean, but prevented normal user accounts from accessing the help.
The first-run experience wasn't very good.