[archive] Update build_date calculation in lastgood query
The lastgood query determines the date of the lastgood pipeline and the
date of the lastgood job within that pipeline. Doing this correctly
requires a two-stage query. First, it identifies the lastgood pipeline
using the selected strategy. Second, it identifies the lastgood job
within the lastgood pipeline.

Jobs are archived if they are older than the lastgood pipeline OR if
they are older than the lastgood job for that pipeline.