Switch to using RefService.FindAllBranches
We were using FindAllBranchNames to load the set of current branches
because all we needed was the branch name to match ci_builds.ref.

To get .gitlab-ci.yml we'll need to use CommitService.TreeEntry which
requires the branch's commit id (revision). Fortunately,
FindAllBranches returns name and commit_id in a single call.

if we want to call BlobService.FindBlob()