Archive artifacts when a build is removed or renamed in the CI config
We archive artifacts by job. A job's artifacts are archived when a newer
instance of that job exists for a given project and branch.

The job names in the CI configuration file gitlab-ci.yml determine the
set of jobs for a pipeline. If a job is renamed or removed from the CI
configuration, jobs with the old name will not be archived, because
there will never be a 'newer' job with that name to trigger archival.
The legacy job is always the LASTGOOD job with that name.

If a job is removed from the CI configuration, it's safe to archive
artifacts for those jobs. It's implied that by updating the CI config we
no longer want or need those older jobs.

NOTE: We never touch tags!