Use project temp table to filter builds
Join the project temp table directly rather than using a project id list
query parameter. This simplifies the interaction with the database and
ensures we won't hit any length limits of parameters or the IN clause.

The join here isn't ideal (creates more rows to aggregate), but I
think it's correct and should be faster than IN (select distinct).