Account for column type differences in the list_projects query
The list_projects recursive CTE uses string concatenation to calculate
the full path to a given namespace. The result type of the concatenation
operation excludes any character limits set on the underlying columns.

On GitLab prior to 8.3 the namespaces.path column is of type
varchar(255) and the concatenation result is of type varchar. This
causes the following error:

   ERROR: recursive query "ns_paths" column 3 has type character
   varying(255) in non-recursive term but type character varying overall
   LINE 3:     select n.id, n.parent_id, n.path
                                      ^
   HINT:  Cast the output of the non-recursive term to the correct type.

To work around this we cast both path columns to the postgres text type.

I tried following the hint and just casting the non-recursive term to
varchar, but the type conversion had no affect on the column's character
limit.

I also tried casting just the non-recursive term to text while
leaving the recursive term as varchar (in an attempt to minimize type
conversions). This worked, but I wasn't sure if it would continue
working in the future and I don't know if there's any measurable cost
associated with the additional cast.

For more information see #13.