Fix get_namespace query when the parent_id is null
When getting a root namespace the project_id parameter is null and the
comparision of the parameter to the projects.id column must use the 'is
null' operator. We tried to handle this case by checking the parameter
for null, but we forgot to ensure the column was also null in this case.
As a result, when searching for a root namespace, ANY namespace with the
same name could be returned.