Handle chunked responses from Gitlay for get_branches
Gitaly "streams" branches 20 at a time. Each gRPC item contains a list
of up to 20 branches. Glartifacts was not aware of this 20-item
chunking, so it was failing for some projects.

This commit iterates the gRPC response and collects the branches from
each page's list.

I don't know how TreeEntry is streamed. We're making a request for a
single blob (.gitlab-ci.yml), so I don't think we need to worry about it.